requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
aptos-sdk==0.4.1
argcomplete==2.0.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.2.0
binary==1.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
chatgpt==2.2212.0
click==8.1.2
contourpy==1.0.7
cryptography==36.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
dj-database-url==0.5.0
Django==4.0.3
django-admin-easy==0.6.1
django-admin-volt==0.0.5
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-grappelli==3.0.3
django-imagekit==4.1.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
django-semantic-admin==0.3.2
djangocms-admin-style==3.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dynaconf==3.1.7
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.89.1
ffmpy==0.3.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gradio==3.16.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hvac==0.11.2
identify==2.5.3
idna==3.3
install==1.3.5
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
knox==0.1.14
linkify-it-py==1.0.3
loguru==0.6.0
markdown-it-py==2.1.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nodeenv==1.7.0
numpy==1.24.1
odfpy==1.4.1
openai==0.26.1
openpyxl==3.0.9
orjson==3.8.5
packaging==23.0
pandas==1.5.2
pdf2image==1.16.2
pilkit==2.0
Pillow==9.0.1
pipx==1.0.0
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rich==13.2.0
shortuuid==1.0.8
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.22.0
tablib==3.2.1
tls-client==0.1.8
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.9
userpath==1.8.0
uvicorn==0.20.0
validators==0.18.2
virtualenv==20.16.3
websockets==10.4
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
